ro-backend/requirements/base.txt
Django==1.6
Pillow==2.3.0
South==0.8.4
dj-database-url==0.2.2
dj-static==0.0.5
django-cors-headers==0.12
django-localflavor==1.0
djangorestframework==2.3.10
gunicorn==18.0
oauthlib==0.6.1
python-openid==2.2.5
python-social-auth==0.1.20
requests==2.2.1
requests-oauthlib==0.4.0
six==1.5.2
static==0.4
wsgiref==0.1.2
